fix: creating indexed account address
