+++ packages/kit-bg/src/states/jotai/atomNames.ts
@@ -28,6 +28,7 @@ export enum EAtomNames {
   v4migrationAtom = 'v4migrationAtom',
   v4migrationPersistAtom = 'v4migrationPersistAtom',
   accountIsAutoCreatingAtom = 'accountIsAutoCreatingAtom',
+  isCreatingIndexedAccountAddressesAtom='isCreatingIndexedAccountAddressesAtom',
   accountManualCreatingAtom = 'accountManualCreatingAtom',
   galleryPersistAtom = 'galleryPersistAtom',
   activeAccountValueAtom = 'activeAccountValueAtom',
